NTXentLoss works with float16, float32, and float64

+++ src/pytorch_metric_learning/reducers/base_reducer.py
@@ -6,7 +6,7 @@ from ..utils.module_with_records import ModuleWithRecords
 class BaseReducer(ModuleWithRecords):
     def forward(self, loss_dict, embeddings, labels):
         self.reset_stats()
-        sub_losses = torch.zeros(len(loss_dict)).to(embeddings.device)
+        sub_losses = torch.zeros(len(loss_dict), dtype=embeddings.dtype).to(embeddings.device)
         loss_count = 0
         for loss_name, loss_info in loss_dict.items():
             self.add_to_recordable_attributes(name=loss_name, is_stat=True)
